requirements.txt
streamlit
fastapi
uvicorn[standard]
psycopg2-binary
python-dotenv
pandas
altair
requests